feat: initial flash implement

+++ keyboards/dummy-kb/src/keymap.rs
@@ -73,7 +73,7 @@ const L4: LayerMap = [
     [ _____ , _____ , _____ , _____ , _____ , _____ , _____ , /**/ _____ , _____ , _____ , _____ , _____ , _____ , _____ ],
 ];
 
-pub const KEYMAP: [Layer; CONFIG.layer_count] = [
+pub const KEYMAP: [Layer; CONFIG.layer_count as usize] = [
     Layer {
         map: L0,
         arrowmouse: false,
